Finished dummy_data.py
Finished script for generating dummy data. Script now generates arcs, lines, ellipses and circles on demand. You can specify the number of elements per image (as a randomly generated between two integers or as a single number when both min and max of the range are the same). Output csv files and tiffs are now timestamped.