Debugging length measurement
An issue with pruning was fixed. Most of the changes are concerning the adaptation of the prune() function to use as a hit-and-miss algorithm for diagonal pixels. Three new functions were added: pixel_length_correction(), define_structure() and find_structure() to help with this.  The full testing script was put in test_fibermorph.py but will be cleaned up after this commit.